refactor: rename useFetchEditorResources file to use-fetch-editor-resources
