+++ src/view/editor/use-fetch-editor-resources.ts
